triage-backend/requirements.txt
# TRIAGE Backend — Pinned Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.35
alembic>=1.13.0
aiosqlite>=0.20.0
redis>=5.1.0
celery[redis]>=5.4.0
transformers>=4.46.0
torch>=2.5.0
datasets>=3.0.0
trl>=0.12.0
peft>=0.13.0
accelerate>=1.0.0
bitsandbytes>=0.44.0
structlog>=24.4.0
tenacity>=9.0.0
pyyaml>=6.0.2
numpy>=1.26.0
httpx>=0.27.0
websockets>=13.0
python-dotenv>=1.0.1
rank-bm25>=0.2.2
openenv-core==0.2.3
pytest-asyncio>=0.24.0
